% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{annotationGR}
\alias{annotationGR}
\alias{annotationGR,RangedAnnotation-method}
\title{Accessor method annotationGR}
\usage{
annotationGR(object)

\S4method{annotationGR}{RangedAnnotation}(object)
}
\arguments{
\item{RangedAnnotation}{}
}
\value{
Slot annotationGR
}
\description{
Accessor method annotationGR
}
\seealso{
\code{\link{GRanges}}
}

