% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{annotationL}
\alias{annotationL}
\alias{annotationL,SequenceBasedAnnotation-method}
\title{Accessor method annotationL}
\usage{
annotationL(object)

\S4method{annotationL}{SequenceBasedAnnotation}(object)
}
\arguments{
\item{SequenceBasedAnnotation}{}
}
\value{
Slot annotationL
}
\description{
Accessor method annotationL
}

