% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{annotationMap}
\alias{annotationMap}
\alias{annotationMap,RangedAnnotation-method}
\title{Accessor method annotationMap}
\usage{
annotationMap(object)

\S4method{annotationMap}{RangedAnnotation}(object)
}
\arguments{
\item{RangedAnnotation}{}
}
\value{
Slot annotationGR
}
\description{
Accessor method annotationMap
}
\seealso{
\code{\link{Hits}}
}

