% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{annotationSummary,EnsemblAnnotation-method}
\alias{annotationSummary}
\alias{annotationSummary,EnsemblAnnotation-method}
\alias{annotationSummary,GRangesBasedAnnotation-method}
\alias{annotationSummary,RefSeqUCSCAnnotation-method}
\title{Summarize annotation}
\usage{
\S4method{annotationSummary}{EnsemblAnnotation}(object, ...)

\S4method{annotationSummary}{GRangesBasedAnnotation}(object, ...)

annotationSummary(object, ...)

\S4method{annotationSummary}{RefSeqUCSCAnnotation}(object, ...)
}
\arguments{
\item{RangedAnnotation}{class}
}
\value{
Annotation data.frames as list
}
\description{
Summarize annotation
}

