% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{annotName}
\alias{annotName}
\alias{annotName,RangedAnnotation-method}
\alias{annotName,SequenceBasedAnnotation-method}
\title{Accessor method annotName}
\usage{
annotName(object)

\S4method{annotName}{RangedAnnotation}(object)

\S4method{annotName}{SequenceBasedAnnotation}(object)
}
\arguments{
\item{RangedAnnotation}{}

\item{SequenceBasedAnnotation}{}
}
\value{
Slot annotationGR

Slot annotName
}
\description{
Accessor method annotName
}

