% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{calculateAlignmentCoverageTwoGRanges}
\alias{calculateAlignmentCoverageTwoGRanges}
\title{Calculates the coverage of two GRanges objects}
\usage{
calculateAlignmentCoverageTwoGRanges(qh, sh)
}
\arguments{
\item{subject}{GRanges object}

\item{query}{Granges object}
}
\value{
data.frame coverage
}
\description{
Calculates the coverage of two GRanges objects
}

