% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{convertRangesToDF}
\alias{convertRangesToDF}
\alias{convertRangesToDF,EnsemblAnnotation-method}
\alias{convertRangesToDF,GRanges-method}
\alias{convertRangesToDF,GRangesBasedAnnotation-method}
\alias{convertRangesToDF,RangedAnnotation-method}
\alias{convertRangesToDF,RefSeqUCSCAnnotation-method}
\title{Convert Ranges object to data frame}
\usage{
convertRangesToDF(object, ...)

\S4method{convertRangesToDF}{RangedAnnotation}(object, ...)

\S4method{convertRangesToDF}{GRangesBasedAnnotation}(object, ...)

\S4method{convertRangesToDF}{GRanges}(object, ...)

\S4method{convertRangesToDF}{RefSeqUCSCAnnotation}(object, ...)

\S4method{convertRangesToDF}{EnsemblAnnotation}(object, ...)
}
\arguments{
\item{RangedAnnotation}{}

\item{GRangesBasedAnnotation}{}

\item{GRanges}{}

\item{RefSeqUCSCAnnotation}{}

\item{EnsemblAnnotation}{}
}
\value{
data.frame

data.frame

data.frame

data.frame

data.frame
}
\description{
Convert Ranges object to data frame
}

