% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{createRObject_gtf}
\alias{createRObject_gtf}
\title{Modifies the gtf file when downloaded from ensembl or ucsc by adding an 'intron' annotation line between exons}
\usage{
createRObject_gtf(pathToGTF, filename, beginMetadata = 6, type = "ucsc")
}
\arguments{
\item{pathToGTF}{the path to the gtf file downloaded from UCSC or ENSEMBL}

\item{filename}{name of the gtf file}

\item{beginMetdata}{The column index indicating when the metadata starts normally column 6 (default)}
}
\value{
RObject for annotation (UCSC, ENSEMBL) as GRanges object
}
\description{
Modifies the gtf file when downloaded from ensembl or ucsc by adding an 'intron' annotation line between exons
}

