% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{EnsemblAnnotation}
\alias{EnsemblAnnotation}
\title{Constructor method for Ranged Annotation class}
\usage{
EnsemblAnnotation(annotName, pathToFile, queryGR)
}
\arguments{
\item{.Object}{GeneAnnotation, GRangesBasedAnnotation}

\item{annotName}{Name of the annotation}

\item{pathToFile}{file path to either an RObject or specific file type see  \code{"rtracklayer"} of the subject annotation}

\item{queryGR}{GRanges object of candidates one wants to annotate}
}
\description{
Constructor method for Ranged Annotation class
}

