% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{executeAnnotation}
\alias{executeAnnotation}
\alias{executeAnnotation,NcbiBlastAnnotation-method}
\title{executeAnnotation}
\usage{
executeAnnotation(object, pathToFile, inputStringSet, ...)

\S4method{executeAnnotation}{NcbiBlastAnnotation}(object, pathToFile,
  inputStringSet, word_size = 11, num_threads = 2, dust = TRUE)
}
\arguments{
\item{object}{currently supported: NcbiBlastAnnotation}

\item{pathToFile}{path to an annotation File (currently ncbi makeblast db formatted database)}

\item{inputStringSet}{XStringSet of the sequences one wants to annotate}

\item{...}{special options for subclass defintions of executeAnnotation}

\item{object}{currently supported: NcbiBlastAnnotation}

\item{pathToFile}{path to an annotation File (currently ncbi makeblast db formatted database)}

\item{inputStringSet}{XStringSet of the sequences one wants to annotate}

\item{ncbiBlast}{specific parameters see: ftp://ftp.ncbi.nlm.nih.gov/blast/executables/blast+/LATEST/}
}
\description{
Executes the annotation based on a XStringSet object

Executes the annotation based on a XStringSet object
}

