% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{feature_annotation_Function_ensembl}
\alias{feature_annotation_Function_ensembl}
\alias{feature_annotation_Function_gr}
\title{feature_Annotation_ensembl helper function}
\usage{
feature_annotation_Function_ensembl(annotGR, annotMap, inputGR, inputIDs)

feature_annotation_Function_gr(annotGR, annotMap, inputGR, inputIDs)
}
\arguments{
\item{annotGR}{}

\item{annotMap}{}

\item{inputGR}{}

\item{inputIDs}{}

\item{annotGR}{}

\item{annotMap}{}

\item{inputGR}{}

\item{inputIDs}{}
}
\value{
data.frame

data.frame
}
\description{
feature_Annotation_ensembl helper function

feature_Annotation_gr helper function
}

