% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getAnnotationByID}
\alias{getAnnotationByID}
\alias{getAnnotationByID,RangedAnnotation-method}
\title{Fetch a specific annoation by ID}
\usage{
getAnnotationByID(object, id, ...)

\S4method{getAnnotationByID}{RangedAnnotation}(object, id)
}
\arguments{
\item{RangedAnnotation}{}
}
\value{
RangedAnnotation
}
\description{
Fetch a specific annoation by ID
}

