% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getFlatTable,EnsemblAnnotation-method}
\alias{getFlatTable}
\alias{getFlatTable,EnsemblAnnotation-method}
\alias{getFlatTable,GRangesBasedAnnotation-method}
\alias{getFlatTable,RefSeqUCSCAnnotation-method}
\title{Get Flat table of annotation for db storage}
\usage{
\S4method{getFlatTable}{EnsemblAnnotation}(object, ...)

\S4method{getFlatTable}{GRangesBasedAnnotation}(object, ...)

getFlatTable(object, ...)

\S4method{getFlatTable}{RefSeqUCSCAnnotation}(object, ...)
}
\arguments{
\item{RangedAnnotation}{class}
}
\value{
Annotation data.frames as list or plain data.frame
}
\description{
Get Flat table of annotation for db storage
}

