% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{initialize,RangedAnnotation-method}
\alias{initialize,RangedAnnotation-method}
\title{Initialisation method for Ranged Annotation class}
\usage{
\S4method{initialize}{RangedAnnotation}(.Object, annotName, pathToFile,
  filename, queryGR, ...)
}
\arguments{
\item{.Object}{EnsemblAnnotation, RefSeqUCSCAnnotation, GRangesBasedAnnotation}

\item{annotName}{Name of the annotation}

\item{pathToFile}{file path to either an RObject or specific file type see  \code{"rtracklayer"} of the subject annotation}

\item{queryGR}{GRanges object of candidates one wants to annotate}

\item{...}{data to include in the new object.  Named arguments
    correspond to slots in the class definition. Unnamed arguments must
    be objects from classes that this class extends.}
}
\description{
Initialisation method for Ranged Annotation class
}

