% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{initialize,SequenceBasedAnnotation-method}
\alias{initialize,SequenceBasedAnnotation-method}
\title{Initialisation method for SequenceBasedAnnotation class}
\usage{
\S4method{initialize}{SequenceBasedAnnotation}(.Object, annotName, pathToFile,
  inputStringSet, ...)
}
\arguments{
\item{.Object}{ncbiBlastAnnotation}

\item{annotName}{of the annotation file}

\item{...}{data to include in the new object.  Named arguments
    correspond to slots in the class definition. Unnamed arguments must
    be objects from classes that this class extends.}

\item{annotationL}{list of annotations}
}
\description{
Initialisation method for SequenceBasedAnnotation class
}

