% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{inputGR}
\alias{inputGR}
\alias{inputGR,RangedAnnotation-method}
\title{Accessor method inputGR}
\usage{
inputGR(object)

\S4method{inputGR}{RangedAnnotation}(object)
}
\arguments{
\item{RangedAnnotation}{}
}
\value{
Slot inputGR
}
\description{
Accessor method inputGR
}
\seealso{
\code{\link{GRanges}}
}

