% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{loadLibraryLocallyAs}
\alias{loadLibraryLocallyAs}
\title{Loads an R object by a given name and returns it}
\usage{
loadLibraryLocallyAs(name, filename)
}
\arguments{
\item{name}{Object name one wants to define}

\item{filename}{Path to file and name}

\item{envir}{environment were the object should be loaded to}
}
\value{
TRUE
}
\description{
Loads an R object by a given name and returns it
}

