% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{NcbiBlastAnnotation-class}
\alias{NcbiBlastAnnotation-class}
\title{NcbiBlastAnnotation Class}
\description{
NcbiBlastAnnotation Class
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{annotName of the annotation \code{"character"}}
   \item{\code{slot2}:}{annotationL Overlap with annotation \code{"list"}}
 }
}

