% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{protein_coding_AnnotationFunction_ensembl}
\alias{protein_coding_AnnotationFunction_ensembl}
\title{protein_coding_Annotation_ensembl helper function}
\usage{
protein_coding_AnnotationFunction_ensembl(annotGR, annotMap, inputGR, inputIDs)
}
\arguments{
\item{annotGR}{}

\item{annotMap}{}

\item{inputGR}{}

\item{inputIDs}{}
}
\value{
data.frame
}
\description{
protein_coding_Annotation_ensembl helper function
}

