% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{RangedAnnotation-class}
\alias{RangedAnnotation-class}
\title{Ranged Annotation Virtual Class, currently superclass of: EnsemblAnnotation, , RangesAnnotation}
\description{
Ranged Annotation Virtual Class, currently superclass of: EnsemblAnnotation, , RangesAnnotation
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{annotName of the annotation \code{"character"}}
   \item{\code{slot2}:}{annotationGR Overlap with annotation \code{"GRanges"}}
   \item{\code{slot3}:}{annotationMap Map between annotation entry and input Entry \code{"Hits"}}
 }
}

