% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{RangedAnnotation}
\alias{RangedAnnotation}
\title{Constructor method for Ranged Annotation class}
\usage{
RangedAnnotation(annotName, pathToFile, filename, queryGR, ...)
}
\arguments{
\item{annotName}{Name of the annotation}

\item{pathToFile}{file path to either an RObject or specific file type see  \code{"rtracklayer"} of the subject annotation}

\item{queryGR}{GRanges object of candidates one wants to annotate}

\item{.Object}{EnsemblAnnotation, RefSeqUCSCAnnotation, GRangesBasedAnnotation}
}
\description{
Constructor method for Ranged Annotation class
}

