% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{RefSeqUCSCAnnotation-class}
\alias{RefSeqUCSCAnnotation-class}
\title{RefSeqUCSCAnnotation extends RangedAnnotation}
\description{
RefSeqUCSCAnnotation extends RangedAnnotation
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{annotName of the annotation \code{"character"}}
   \item{\code{slot2}:}{annotationGR Overlap with annotation \code{"GRanges"}}
   \item{\code{slot3}:}{annotationMap Map between annotation entry and input Entry \code{"Hits"}}
   \item{\code{slot4}:}{inputGR Input of user that got annotated \code{"GRanges"}}
 }
}

