% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{RefSeqUCSCAnnotation}
\alias{RefSeqUCSCAnnotation}
\title{Constructor method for RefSeqUCSCAnnotation Annotation class}
\usage{
RefSeqUCSCAnnotation(annotName, pathToFile, filename, queryGR)
}
\arguments{
\item{.Object}{RefSeqUCSCAnnotation, GRangesBasedAnnotation}

\item{annotName}{Name of the annotation}

\item{pathToFile}{file path to either an RObject or specific file type see  \code{"rtracklayer"} of the subject annotation}

\item{queryGR}{GRanges object of candidates one wants to annotate}
}
\description{
Constructor method for RefSeqUCSCAnnotation Annotation class
}

