% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{SequenceBasedAnnotation-class}
\alias{SequenceBasedAnnotation-class}
\title{SequenceBasedAnnotation Virtual Class, currently superclass of: ncbiBlastAnnotation}
\description{
SequenceBasedAnnotation Virtual Class, currently superclass of: ncbiBlastAnnotation
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{annotName of the annotation \code{"character"}}
   \item{\code{slot2}:}{annotationL Overlap with annotation \code{"list"}}
 }
}

