% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{importRangedAnnotation,EnsemblAnnotation-method}
\alias{importRangedAnnotation}
\alias{importRangedAnnotation,EnsemblAnnotation-method}
\alias{importRangedAnnotation,GRangesBasedAnnotation-method}
\alias{importRangedAnnotation,RefSeqUCSCAnnotation-method}
\title{importRangedAnnotation}
\usage{
\S4method{importRangedAnnotation}{EnsemblAnnotation}(object, pathToFile,
  filename)

\S4method{importRangedAnnotation}{GRangesBasedAnnotation}(object, pathToFile,
  filename)

importRangedAnnotation(object, pathToFile, filename, ...)

\S4method{importRangedAnnotation}{RefSeqUCSCAnnotation}(object, pathToFile,
  filename)
}
\arguments{
\item{.object}{EnsemblAnnotation}

\item{pathToFile}{path to a gtf file or .rda file}

\item{filename}{}

\item{.object}{GRangesBasedAnnotation}

\item{.object}{RefSeqUCSCAnnotation}
}
\value{
Ensembl annotation object rtracklayer/GRanges

pathToFile Path to a file supported by rtracklayer/GRanges

pathToFile Path to a file supported by rtracklayer/GRanges
}
\description{
importRangedAnnotation

Import RangedAnnotation
}

