% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{importRangedAnnotation,EnsemblAnnotation-method}
\alias{EnsemblAnnotation}
\alias{importRangedAnnotation}
\alias{importRangedAnnotation,}
\alias{importRangedAnnotation,EnsemblAnnotation-method}
\alias{importRangedAnnotation,GRangesBasedAnnotation-method}
\alias{importRangedAnnotation,RefSeqUCSCAnnotation-method}
\title{Import Range dAnnotation}
\usage{
\S4method{importRangedAnnotation}{EnsemblAnnotation}(object, pathToFile,
  filename, ...)

\S4method{importRangedAnnotation}{GRangesBasedAnnotation}(object, pathToFile,
  filename, ...)

importRangedAnnotation(object, pathToFile, filename, ...)

\S4method{importRangedAnnotation}{RefSeqUCSCAnnotation}(object, pathToFile,
  filename, ...)
}
\arguments{
\item{Object}{}

\item{pathToFile}{path to a gtf or rda file}

\item{filename}{name of the gtf/rda file}
}
\description{
Import Range dAnnotation
}

